Deleting a non existant registration now returns 404

+++ src/main/java/com/hackathlon/hackathlon/service/impl/RegistrationServiceImpl.java
@@ -73,6 +73,9 @@ public class RegistrationServiceImpl implements RegistrationService {
 
     @Override
     public void delete(Registration registration) {
+        if (!this.registrationRepository.existsById(registration.getID())) {
+            throw new NoSuchElementException("Registration with id " + registration.getID() + " not found");
+        }
         this.registrationRepository.deleteById(registration.getID());
 //        this.registrationRepository.delete(registration);
     }
